+++ src/algorand/AlgorandEngine.ts
@@ -461,12 +461,7 @@ export class AlgorandEngine extends CurrencyEngine<
       this.tokenCheckTransactionsStatus[cc] = 1
     }
     this.updateOnAddressesChecked()
-
-    if (this.transactionEvents.length > 0) {
-      this.walletLocalDataDirty = true
-      this.currencyEngineCallbacks.onTransactions(this.transactionEvents)
-      this.transactionEvents = []
-    }
+    this.updateTransactionEvents()
   }
 
   // // ****************************************************************************