+++ src/eos/EosEngine.ts
@@ -540,10 +540,7 @@ export class EosEngine extends CurrencyEngine<EosTools, SafeEosWalletInfo> {
 
       this.tokenCheckTransactionsStatus[token] = 1
       this.updateOnAddressesChecked()
-      if (this.transactionEvents.length > 0) {
-        this.currencyEngineCallbacks.onTransactions(this.transactionEvents)
-        this.transactionEvents = []
-      }
+      this.updateTransactionEvents()
     }
   }